separate app.py from functions.py
